requirements/requirements.txt
alembic==1.10.4
anyio==3.6.2
asyncio==3.4.3
asyncpg==0.27.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.11
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
